sales_agent/requirements.txt
python-dotenv
langchain
sentence-transformers